cleanup gui files
